ConfigLoader ensures download folder exists.
We manage it here so that each user who makes a checkout gets
a folder that they can use, rather than just having one folder
for all admin users.

The alternative is to create the appropriate folder on login,
but that scatters the logic around the codebase in an unpleasant
way.